Kernel params set by the config daemon should be verified
In the case of vfio-pci, the config daemon might not notice it has to
reboot to add kernel parameters. This could happen if the vfio driver
state was updated but the config daemon was interrupted (e.g policy was
removed and readded quickly).

This could lead to a pending change to the kernel parameters via grubby
or os-tree but without a reboot, these changes wouldn't be applied.

This change introduces a map to hold the desired kernel parameters that
the config daemon wishes to apply. This map is then validated against
"/proc/cmdline" to see if the desired kernel parameters are actually
applied. If not then there will be a subsequent attempt to apply these
parameters.

We need to make sure that the kernel parameters are set, thus we should
pass the errors up to the config daemon instead of silently consuming
the errors when setting kernel parameters.

Signed-off-by: William Zhao <wizhao@redhat.com>